requirements.txt
python-dotenv~=1.0.0
flask~=3.0.3
Flask-SocketIO~=5.3.6
werkzeug==3.0.6
simple-websocket==1.0.0
python-socketio==5.11.3
python-engineio==4.9.1
flask-session==0.8.0
opencv-python~=4.8.0.74
